added draft documents export
